+++ continuousintegration/docker/ctafrontend/opt/run/bin/taped.sh
@@ -42,19 +42,22 @@ echo ${DATABASEURL} > /etc/cta/cta-catalogue.conf
 TAPED_CONF_FILE="/etc/cta/cta-taped-${DRIVENAMES[${driveslot}]}.conf"
 
 # cta-taped setup
-  echo "taped BufferSizeBytes 262144" > "${TAPED_CONF_FILE}"
-  echo "taped BufferCount 200" >> "${TAPED_CONF_FILE}"
-  echo "taped MountCriteria 2000000, 100" >> "${TAPED_CONF_FILE}"
-  echo "taped WatchdogIdleSessionTimer 2" >> "${TAPED_CONF_FILE}" # Make tape servers more responsive, thus improving CI test speed
-  echo "ObjectStore BackendPath $OBJECTSTOREURL" >> "${TAPED_CONF_FILE}"
-  echo "taped UseEncryption no" >> "${TAPED_CONF_FILE}"
-  echo "taped DriveName ${DRIVENAMES[${driveslot}]}" >> "${TAPED_CONF_FILE}"
-  echo "taped DriveLogicalLibrary ${DRIVENAMES[${driveslot}]}" >> "${TAPED_CONF_FILE}"
-  echo "taped DriveDevice /dev/${DRIVEDEVICES[${driveslot}]}" >> "${TAPED_CONF_FILE}"
-  echo "taped DriveControlPath smc${driveslot}" >> "${TAPED_CONF_FILE}"
+echo "taped BufferSizeBytes 262144" > "${TAPED_CONF_FILE}"
+echo "taped BufferCount 200" >> "${TAPED_CONF_FILE}"
+echo "taped MountCriteria 2000000, 100" >> "${TAPED_CONF_FILE}"
+echo "taped WatchdogIdleSessionTimer 2" >> "${TAPED_CONF_FILE}" # Make tape servers more responsive, thus improving CI test speed
+echo "ObjectStore BackendPath $OBJECTSTOREURL" >> "${TAPED_CONF_FILE}"
+echo "taped UseEncryption no" >> "${TAPED_CONF_FILE}"
+echo "taped DriveName ${DRIVENAMES[${driveslot}]}" >> "${TAPED_CONF_FILE}"
+echo "taped DriveLogicalLibrary ${DRIVENAMES[${driveslot}]}" >> "${TAPED_CONF_FILE}"
+echo "taped DriveDevice /dev/${DRIVEDEVICES[${driveslot}]}" >> "${TAPED_CONF_FILE}"
+echo "taped DriveControlPath smc${driveslot}" >> "${TAPED_CONF_FILE}"
+# Decrease schedulerDB cache timeout for tests
+echo "taped TapeCacheMaxAgeSecs 1" >> "${TAPED_CONF_FILE}"
+echo "taped RetrieveQueueCacheMaxAgeSecs 1" >> "${TAPED_CONF_FILE}"
 
-  echo "general InstanceName CI" >> "${TAPED_CONF_FILE}"
-  echo "general SchedulerBackendName VFS" >> "${TAPED_CONF_FILE}"
+echo "general InstanceName CI" >> "${TAPED_CONF_FILE}"
+echo "general SchedulerBackendName VFS" >> "${TAPED_CONF_FILE}"
 
 
 ####