+++ objectstore/ObjectStoreFixture.cpp
@@ -23,7 +23,7 @@ namespace unitTests {
 
   void ObjectStore::SetUp() {
     // We need to cleanup the queue statistics cache before every test
-    cta::objectstore::Helpers::flushRetrieveQueueStatisticsCache();
+    cta::objectstore::Helpers::flushStatisticsCache();
   }
 
 }