+++ scheduler/rdbms/RelationalDB.hpp
@@ -119,7 +119,9 @@ class RelationalDB: public SchedulerDatabase {
     const cta::common::dataStructures::RetrieveFileQueueCriteria &criteria, const std::optional<std::string> diskSystemName,
     log::LogContext &logContext) override;
 
-  void clearRetrieveQueueStatisticsCache(const std::string & vid) override;
+  void clearStatisticsCache(const std::string & vid) override;
+
+  void setStatisticsCacheConfig(const StatisticsCacheConfig & conf) override;
 
   void cancelRetrieve(const std::string& instanceName, const cta::common::dataStructures::CancelRetrieveRequest& rqst,
     log::LogContext& lc) override;